# docs/swagger.yaml
basePath: /
definitions:
  models.Admin:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
    type: object
  models.AdminCreateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.AdminDataResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.AdminUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.Dosen:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      image:
        type: string
      jabatan:
        type: string
      kapasitas:
        type: integer
      kepakaran:
        type: string
      name:
        type: string
      nidn:
        type: string
      password:
        type: string
      prodi:
        type: string
      total_mahasiswa:
        type: integer
      updatedAt:
        type: string
    type: object
  models.DosenCreateRequest:
    properties:
      email:
        type: string
      image:
        type: string
      jabatan:
        type: string
      kapasitas:
        type: integer
      kepakaran:
        type: string
      name:
        type: string
      nidn:
        type: string
      password:
        type: string
      prodi:
        type: string
      total_mahasiswa:
        type: integer
    type: object
  models.DosenDataResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      image:
        type: string
      jabatan:
        type: string
      kapasitas:
        type: integer
      kepakaran:
        type: string
      name:
        type: string
      nidn:
        type: string
      prodi:
        type: string
      total_mahasiswa:
        type: integer
      updatedAt:
        type: string
    type: object
  models.DosenUpdateRequest:
    properties:
      email:
        type: string
      image:
        type: string
      jabatan:
        type: string
      kapasitas:
        type: integer
      kepakaran:
        type: string
      name:
        type: string
      nidn:
        type: string
      password:
        type: string
      prodi:
        type: string
      total_mahasiswa:
        type: integer
    type: object
...
host: projectskripsi-fvwdncsc.b4a.run
info:
  contact: {}
  description: An API for Sijudul App using Gin
  title: Sijudul API
  version: "1.0"
paths:
  /admin:
    get:
      description: Get All Admins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AdminDataResponse'
            type: array
      summary: Get All Admin
      tags:
      - Admin
    post:
      description: Create Admins
      parameters:
      - description: Raw Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AdminCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
      summary: Create Admin
      tags:
      - Admin
  /admin/{id}:
    delete:
      description: Delete Admins
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete Admin
      tags:
      - Admin
    get:
      description: Get Admin By ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminDataResponse'
      summary: Get Admin By ID
      tags:
      - Admin
    put:
      description: Update Admins
      parameters:
      - description: Raw Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AdminUpdateRequest'
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
      summary: Update Admin
      tags:
      - Admin
  /dosen:
    get:
      description: Get All Dosens
      parameters:
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DosenDataResponse'
            type: array
      summary: Get All Dosen
      tags:
      - Dosen
    post:
      description: Create Dosen
      parameters:
      - description: Raw Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DosenCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Dosen'
      summary: Create Dosen
      tags:
      - Dosen
  /dosen/{id}:
    delete:
      description: Delete Dosen
      parameters:
      - description: Dosen ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete Dosen
      tags:
      - Dosen
    get:
      description: Get Dosen By ID
      parameters:
      - description: Dosen ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DosenDataResponse'
      summary: Get Dosen By ID
      tags:
      - Dosen
    put:
      description: Update Dosen
      parameters:
      - description: Raw Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DosenUpdateRequest'
      - description: Dosen ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Dosen'
      summary: Update Dosen
      tags:
      - Dosen
...
schemes:
- https
- http
swagger: "2.0"